requirements.txt
asgiref==3.4.1
async-generator==1.10
attrs==21.2.0
beautifulsoup4==4.10.0
boto3==1.24.62
botocore==1.27.62
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.7
confuse==2.0.0
cryptography==35.0.0
Django==3.2.9
django-environ==0.8.1
django-widget-tweaks==1.4.12
django-widgets-improved==1.5.0
ebaysdk==2.2.0
gdata==2.0.18
google==3.0.0
google-api-core==2.2.2
google-api-python-client==2.29.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.53.0
googletrans==3.0.0
gspread==4.0.1
gunicorn==20.1.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.11.1
httpcore==0.9.1
httplib2==0.20.2
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
jmespath==1.0.1
lxml==4.6.4
mysqlclient==2.1.1
oauth2client==4.1.3
oauthlib==3.1.1
outcome==1.1.0
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDrive==1.3.1
pygsheets==2.0.5
pyOpenSSL==21.0.0
pyparsing==3.0.4
PySocks==1.7.1
python-amazon-sp-api==0.16.0
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2021.3
PyYAML==6.0
requests==2.26.0
requests-oauthlib==1.3.0
rfc3986==1.5.0
rsa==4.7.2
s3transfer==0.6.0
selenium==4.0.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3
sqlparse==0.4.2
tbselenium==0.5.3
tqdm==4.64.0
trio==0.19.0
trio-websocket==0.9.2
uritemplate==4.1.1
urllib3==1.26.7
webdriver-manager==3.8.3
wsproto==1.0.0